# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for IPL Backend Service with menu management",
        "title": "IPL Backend Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/billings/bulk-custom": {
            "post": {
                "description": "Create custom billings for specified user IDs or all penghuni users if user_ids is empty. Requires auth-token cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billings"
                ],
                "summary": "Create bulk custom billings",
                "parameters": [
                    {
                        "description": "Bulk billing request with month and year",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.BulkBillingCustomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bulk billing creation result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.BulkBillingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/billings/bulk-monthly": {
            "post": {
                "description": "Create monthly billings for specified user IDs or all penghuni users if user_ids is empty. Requires auth-token cookie.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billings"
                ],
                "summary": "Create bulk monthly billings",
                "parameters": [
                    {
                        "description": "Bulk billing request with month and year",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.BulkBillingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bulk billing creation result",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/service.BulkBillingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/billings/confirm-payment": {
            "post": {
                "description": "Receive payment gateway webhook and process payment confirmation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billings"
                ],
                "summary": "Confirm payment webhook",
                "parameters": [
                    {
                        "description": "Webhook payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ConfirmPaymentWebhookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Webhook received",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/billings/confirm-single": {
            "post": {
                "description": "Confirm payment by sending a single billing_id in JSON body",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billings"
                ],
                "summary": "Confirm single billing payment",
                "parameters": [
                    {
                        "description": "Billing ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ConfirmPaymentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment confirmed",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid payload",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/billings/penghuni": {
            "get": {
                "description": "Retrieve all billing penghuni records without pagination or search",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billings"
                ],
                "summary": "Get all billing penghuni",
                "responses": {
                    "200": {
                        "description": "Billing penghuni retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.BillingPenghuniResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/billings/penghuni/search": {
            "get": {
                "description": "Get billing data for penghuni users. Supports pagination and search by `q` (nama_penghuni or user ID).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "billings"
                ],
                "summary": "Get billing penghuni list with summed nominals",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search by nama_penghuni or ID",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Billing penghuni retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/utils.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.BillingPenghuniResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.BulkBillingCustomRequest": {
            "type": "object",
            "required": [
                "billing_settings_id",
                "month",
                "year"
            ],
            "properties": {
                "billing_settings_id": {
                    "description": "Billing settings ID",
                    "type": "integer"
                },
                "month": {
                    "description": "Month 1-12",
                    "type": "integer",
                    "maximum": 12,
                    "minimum": 1
                },
                "user_ids": {
                    "description": "Empty means all penghuni users",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "year": {
                    "description": "Reasonable year range",
                    "type": "integer",
                    "maximum": 2100,
                    "minimum": 2020
                }
            }
        },
...
        "handler.ConfirmPaymentRequest": {
            "type": "object",
            "required": [
                "billing_id"
            ],
            "properties": {
                "billing_id": {
                    "type": "integer",
                    "example": 123
                }
            }
        },
...
        "service.BulkBillingResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "failed_count": {
                    "type": "integer"
                },
                "success_count": {
                    "type": "integer"
                },
                "total_billings": {
                    "type": "integer"
                },
                "total_users": {
                    "type": "integer"
                }
            }
        },
...
        "utils.APIResponse": {
            "description": "Standard API response structure",
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    }
}